QE-68: Move ProjectUtils from 'io.quarkus.eclipse.core' to 'io.quarkus.eclipe.core.project'
