+++ plugin/io.quarkus.eclipse.core/META-INF/MANIFEST.MF
@@ -11,4 +11,4 @@ Require-Bundle: org.eclipse.core.resources,
  org.eclipse.core.runtime,
  org.eclipse.m2e.core,
  io.quarkus.eclipse.runtime
-Export-Package: io.quarkus.eclipse.core
+Export-Package: io.quarkus.eclipse.core.project
