+++ test/io.quarkus.eclipse.core.test/src/io/quarkus/eclipse/core/project/ProjectUtilsTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.quarkus.eclipse.core;
+package io.quarkus.eclipse.core.project;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -27,6 +27,8 @@ import org.eclipse.core.runtime.CoreException;
 import org.eclipse.core.runtime.NullProgressMonitor;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+
+import io.quarkus.eclipse.core.project.ProjectUtils;
 
 public class ProjectUtilsTest {
